added three unit test cases to travel_time under project_tests/assets
