+++ project/logs/travel_time_1717010012.9036608.log
@@ -1,0 +1,7 @@
+2024-05-29 15:13:33,122 - travel_time - INFO - Starting pipeline run
+2024-05-29 15:13:33,122 - travel_time - INFO - Getting pipeline environment variables
+2024-05-29 15:13:33,122 - travel_time - INFO - Creating Travel Time API client
+2024-05-29 15:13:33,122 - travel_time - INFO - Extracting data from Travel Time API
+2024-05-29 15:13:35,939 - travel_time - INFO - Adding load_timestamp and load_id columns to dataframe
+2024-05-29 15:13:35,939 - travel_time - INFO - Loading data to postgres
+2024-05-29 15:13:36,071 - travel_time - INFO - Pipeline run successful
