+++ project/logs/travel_time_1717010049.20877.log
@@ -1,0 +1,7 @@
+2024-05-29 15:14:09,412 - travel_time - INFO - Starting pipeline run
+2024-05-29 15:14:09,413 - travel_time - INFO - Getting pipeline environment variables
+2024-05-29 15:14:09,413 - travel_time - INFO - Creating Travel Time API client
+2024-05-29 15:14:09,413 - travel_time - INFO - Extracting data from Travel Time API
+2024-05-29 15:14:12,182 - travel_time - INFO - Adding load_timestamp and load_id columns to dataframe
+2024-05-29 15:14:12,206 - travel_time - INFO - Loading data to postgres
+2024-05-29 15:14:12,356 - travel_time - INFO - Pipeline run successful
